Refactor Dialog creation
